CRUD Finalizado

+++ Domain/Entidades/Produto.cs
@@ -4,7 +4,7 @@
     {
         public int IdProduto { get; set; }
         public string NomeProduto { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
         public string Descricao { get; set; }
     }
 }
